Delete a bunch of commented-out code.
