+++ example/src/app/controllers/organizations.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"github.com/68696c6c/goat"
 	"github.com/68696c6c/goat/controller"
-	"github.com/68696c6c/goat/query2"
+	"github.com/68696c6c/goat/query"
 	"github.com/68696c6c/goat/resource"
 	"github.com/gin-gonic/gin"
 
@@ -16,48 +16,18 @@ type OrganizationsController interface {
 }
 
 type organizations struct {
-	// baseUrl string
-	// baseUrl *url.URL
-	// errors goat.ErrorHandler
 	repo repos.OrganizationsRepo
 }
 
-func NewOrganizationsController(
-	// baseUrl *url.URL,
-	// baseUrl string,
-	// errors goat.ErrorHandler,
-	repo repos.OrganizationsRepo,
-) OrganizationsController {
+func NewOrganizationsController(repo repos.OrganizationsRepo) OrganizationsController {
 	return organizations{
-		// baseUrl: baseUrl,
-		// baseUrl: baseUrl,
-		// errors: errors,
 		repo: repo,
 	}
 }
 
-// func (c organizations) GetErrorHandler() goat.ErrorHandler {
-// 	return c.errors
-// }
-
-// func (c organizations) GetLink(path string) resource.Link {
-// 	println("ORGANIZATIONS GET LINK")
-// 	return goat.GetLink(models.OrganizationLinkKey, path)
-// 	// return resource.MakeLink(c.baseUrl.JoinPath(path).String())
-// }
-
-// func (c organizations) GetBaseUrl() string {
-// 	return c.baseUrl
-// }
-
-// func (c organizations) GetBaseUrl() *url.URL {
-// 	return goat.GetUrl(models.OrganizationLinkKey)
-// }
-
 func (c organizations) List(cx *gin.Context) {
-	// controller.HandleList[*models.Organization](cx, c, c.repo)
 	queryString := cx.Request.URL.Query()
-	q := query2.NewQueryFromUrl(queryString)
+	q := query.NewQueryFromUrl(queryString)
 	p := resource.NewPaginationFromUrl(queryString)
 	controller.HandleList[*models.Organization](cx, c.repo, goat.GetUrl(models.OrganizationLinkKey), q, p)
 }
