+++ example/src/app/repos/organizations.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/68696c6c/goat"
-	"github.com/68696c6c/goat/query2"
+	"github.com/68696c6c/goat/query"
 	"github.com/68696c6c/goat/repo"
 	"github.com/68696c6c/goat/resource"
 	"gorm.io/gorm"
@@ -33,7 +33,6 @@ func (r organizationsRepo) Make() *models.Organization {
 
 func (r organizationsRepo) Create(_ context.Context, u models.OrganizationRequest) (*models.Organization, error) {
 	m := r.Make()
-	// m.ParentId = *u.ParentId
 	m.Name = *u.Name
 	m.Website = *u.Website
 	return m, nil
@@ -44,7 +43,6 @@ func (r organizationsRepo) Update(cx context.Context, id goat.ID, u models.Organ
 	if err != nil {
 		return nil, err
 	}
-	// m.ParentId = goat.ValueOrDefault[goat.ID](u.ParentId, m.ParentId)
 	m.Name = goat.ValueOrDefault[string](u.Name, m.Name)
 	m.Website = goat.ValueOrDefault[string](u.Website, m.Website)
 	return m, nil
@@ -54,11 +52,8 @@ func (r organizationsRepo) getBaseQuery() *gorm.DB {
 	return r.db.Model(&models.Organization{})
 }
 
-// func (r organizationsRepo) Filter(cx context.Context, p resource.Pagination, filters ...repo.QueryFilter) ([]*models.Organization, resource.Pagination, error) {
-func (r organizationsRepo) Filter(cx context.Context, q query2.Builder, p resource.Pagination) ([]*models.Organization, resource.Pagination, error) {
+func (r organizationsRepo) Filter(cx context.Context, q query.Builder, p resource.Pagination) ([]*models.Organization, resource.Pagination, error) {
 	base := r.db.WithContext(cx).Model(&models.Organization{})
-	// return repo.PaginatedFilter[models.Organization](base, p, filters...)
-	// return repo.FilterTemp[models.Organization](q, r.db)
 	return repo.Filter[models.Organization](base, q, p)
 }
 
@@ -81,9 +76,3 @@ func (r organizationsRepo) Save(cx context.Context, m *models.Organization) erro
 func (r organizationsRepo) Delete(cx context.Context, m *models.Organization) error {
 	return repo.Delete[*models.Organization](r.db.WithContext(cx), m)
 }
-
-// func (r organizationsRepo) GetChildren(cx context.Context, orgId goat.ID, q query2.Builder, p resource.Pagination) ([]*models.Organization, resource.Pagination, error) {
-// 	base := r.db.WithContext(cx).Model(&models.Organization{})
-// 	q.Where("parent_id", query2.Equals, orgId)
-// 	return repo.Filter[models.Organization](base, q, p)
-// }
