+++ example/src/app/repos/users.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/68696c6c/goat"
-	"github.com/68696c6c/goat/query2"
+	"github.com/68696c6c/goat/query"
 	"github.com/68696c6c/goat/repo"
 	"github.com/68696c6c/goat/resource"
 	"gorm.io/gorm"
@@ -17,12 +17,8 @@ import (
 
 type UsersRepo interface {
 	repo.CRUD[*models.User, models.UserRequest]
-	// ValidateCreate(validator validator.StructLevel)
-	ApplyFilterForUser(q query2.Builder, user *models.User) error
-	FilterStrings(query query2.Builder, fields map[string][]string) error
-	// UserLevelFilter(user models.User) repo.QueryFilter
-	// SearchFilter(fields url.Values) repo.QueryFilter
-	// FilterTemp(cx context.Context, p query.Pagination, scopes ...repo.QueryScope) ([]*models.User, query.Pagination, error)
+	ApplyFilterForUser(q query.Builder, user *models.User) error
+	FilterStrings(query query.Builder, fields map[string][]string) error
 }
 
 type usersRepo struct {
@@ -38,32 +34,6 @@ func NewUsersRepo(db *gorm.DB) UsersRepo {
 func (r usersRepo) Make() *models.User {
 	return models.MakeUser()
 }
-
-// func (r usersRepo) ValidateCreate(validator validator.StructLevel) {
-// 	req := validator.Current().Interface().(models.UserRequest)
-//
-// 	if req.Email == nil {
-// 		validator.ReportError(req.Email, "email", "", "required", "")
-// 	}
-//
-// 	_, err := repo.First[*models.User](r.db, &models.User{}, "email = ?", req.Email)
-// 	if !goat.RecordNotFound(err) {
-// 		validator.ReportError(req.Email, "email", "", "unique", "")
-// 	}
-// }
-//
-// func (r usersRepo) ValidateUpdate(validator validator.StructLevel) {
-// 	req := validator.Current().Interface().(models.UserRequest)
-//
-// 	if req.Email == nil {
-// 		validator.ReportError(req.Email, "email", "", "required", "")
-// 	}
-//
-// 	_, err := repo.First[*models.User](r.db, &models.User{}, "email = ?", req.Email)
-// 	if !goat.RecordNotFound(err) {
-// 		validator.ReportError(req.Email, "email", "", "unique", "")
-// 	}
-// }
 
 func (r usersRepo) Create(cx context.Context, u models.UserRequest) (*models.User, error) {
 	m := r.Make()
@@ -115,58 +85,24 @@ func (r usersRepo) getBaseQuery() *gorm.DB {
 	return r.db.Model(&models.User{})
 }
 
-// func (r usersRepo) Filter(cx context.Context, q query.Builder) ([]*models.User, error) {
-//
-// 	return repo.FilterTemp[models.User](q, r.db.WithContext(cx).Model(&models.User{}))
-// 	// return repo.Filter[models.User](q, r.getBaseQuery)
-// }
-
-// func (r usersRepo) Filter(cx context.Context, p resource.Pagination, filters ...repo.QueryFilter) ([]*models.User, resource.Pagination, error) {
-func (r usersRepo) Filter(cx context.Context, q query2.Builder, p resource.Pagination) ([]*models.User, resource.Pagination, error) {
+func (r usersRepo) Filter(cx context.Context, q query.Builder, p resource.Pagination) ([]*models.User, resource.Pagination, error) {
 	base := r.db.WithContext(cx).Model(&models.User{})
 	return repo.Filter[models.User](base, q, p)
-	// // for _, scope := range scopes {
-	// // 	base.Scopes(scope)
-	// // }
-	// result, pagination, err := repo.PaginatedFilter[models.User](base, p, filters...)
-	// if err != nil {
-	// 	return result, pagination, err
-	// }
-	// // pp := query.NewPaginationFromValues(p.Page, p.PageSize, count)
-	// // return repo.Filter[models.User](q, r.getBaseQuery)
-	// return result, pagination, nil
 }
 
-// func (r usersRepo) UserLevelFilter(user models.User) repo.QueryFilter {
-// 	return func(db *gorm.DB) *gorm.DB {
-// 		switch user.Level {
-// 		case enums.UserLevelSuper:
-// 			// Super-users can see everything.
-// 			break
-// 		case enums.UserLevelCustomer:
-// 			// Customers can only see themselves.
-// 			db.Where("id = ?", user.ID)
-// 		default:
-// 			// Admin and normal users can only see users in their own org.
-// 			db.Where("organization_id = ?", user.OrganizationId)
-// 		}
-// 		return db
-// 	}
-// }
-
-func (r usersRepo) ApplyFilterForUser(q query2.Builder, user *models.User) error {
+func (r usersRepo) ApplyFilterForUser(q query.Builder, user *models.User) error {
 	switch user.Level {
 	case enums.UserLevelSuper:
 		// Super-users can see everything.
 		break
 	case enums.UserLevelCustomer:
 		// Customers can only see themselves.
-		// q.WhereEq("id", user.ID)
-		q.Where("id", query2.Equals, user.ID)
+		q.WhereEq("id", user.ID)
+		// q.Where("id", query2.Equals, user.ID)
 	default:
 		// Admin and normal users can only see users in their own org.
-		// q.WhereEq("organization_id", user.OrganizationId)
-		q.Where("organization_id", query2.Equals, user.OrganizationId)
+		q.WhereEq("organization_id", user.OrganizationId)
+		// q.Where("organization_id", query2.Equals, user.OrganizationId)
 	}
 	return nil
 }
@@ -191,139 +127,17 @@ func (r usersRepo) Delete(cx context.Context, m *models.User) error {
 	return repo.Delete[*models.User](r.db.WithContext(cx), m)
 }
 
-func (r usersRepo) FilterStrings(q query2.Builder, fields map[string][]string) error {
+func (r usersRepo) FilterStrings(q query.Builder, fields map[string][]string) error {
 	for fieldName, values := range fields {
 		switch fieldName {
 		case "email":
 			fallthrough
 		case "name":
 			if len(values) > 0 {
-				// q.WhereLike(fieldName, fmt.Sprintf("%%%s%%", values[0]))
-				q.Where(fieldName, query2.Like, fmt.Sprintf("%%%s%%", values[0]))
+				q.WhereLike(fieldName, fmt.Sprintf("%%%s%%", values[0]))
+				// q.Where(fieldName, query2.Like, fmt.Sprintf("%%%s%%", values[0]))
 			}
 		}
 	}
 	return nil
 }
-
-// func filterStrings(db *gorm.DB, fields url.Values) *gorm.DB {
-// 	for fieldName, values := range fields {
-// 		switch fieldName {
-// 		case "email":
-// 			fallthrough
-// 		case "name":
-// 			if len(values) > 0 {
-// 				cond := fmt.Sprintf("%s LIKE ?", fieldName)
-// 				db.Or(cond, fmt.Sprintf("%%%s%%", values[0]))
-// 				// f.WhereField(fieldName, filter.OpLike, fmt.Sprintf("%%%s%%", values[0]))
-// 			}
-// 		}
-// 	}
-// 	return db
-// }
-//
-// func (r usersRepo) SearchFilter(fields url.Values) repo.QueryFilter {
-// 	return func(db *gorm.DB) *gorm.DB {
-// 		if len(fields) == 0 {
-// 			return db
-// 		}
-// 		return db.Where(filterStrings(db, fields))
-// 	}
-// }
-//
-// type tempFilter struct {
-// 	condition string
-// 	value     string
-// }
-//
-// func (r usersRepo) filterStrings(db *gorm.DB, fields url.Values) *gorm.DB {
-// 	// var result []tempFilter
-// 	// q := db.Unscoped()
-// 	var set bool
-// 	for fieldName, values := range fields {
-// 		switch fieldName {
-// 		case "email":
-// 			fallthrough
-// 		case "name":
-// 			if len(values) > 0 {
-// 				// result = append(result, tempFilter{
-// 				// 	condition: fmt.Sprintf("%s LIKE ?", fieldName),
-// 				// 	value:     fmt.Sprintf("%%%s%%", values[0]),
-// 				// })
-// 				cond := fmt.Sprintf("%s LIKE ?", fieldName)
-// 				value := fmt.Sprintf("%%%s%%", values[0])
-// 				if set {
-// 					db.Or(cond, value)
-// 				} else {
-// 					set = true
-// 					db.Where(cond, value)
-// 				}
-// 				// // f.WhereField(fieldName, filter.OpLike, fmt.Sprintf("%%%s%%", values[0]))
-// 			}
-// 		}
-// 	}
-// 	return db
-// }
-//
-// func (r usersRepo) filterStrings(fields url.Values) repo.QueryFilter {
-// 	return func(db *gorm.DB) *gorm.DB {
-// 		// db.Where("1 = 1")
-// 		for fieldName, values := range fields {
-// 			switch fieldName {
-// 			case "email":
-// 				fallthrough
-// 			case "name":
-// 				if len(values) > 0 {
-// 					cond := fmt.Sprintf("%s LIKE ?", fieldName)
-// 					db.Or(cond, fmt.Sprintf("%%%s%%", values[0]))
-// 					// f.WhereField(fieldName, filter.OpLike, fmt.Sprintf("%%%s%%", values[0]))
-// 				}
-// 			}
-// 		}
-// 		return db
-// 		// where, params, err := f.Apply()
-// 		// if err != nil {
-// 		// 	// return nil, errors.Wrap(err, "failed to apply filter")
-// 		// }
-// 		// var filters []filter
-// 		// db.Where("1 = 1")
-// 		// for fieldName, values := range fields {
-// 		// 	switch fieldName {
-// 		// 	case "email":
-// 		// 		fallthrough
-// 		// 	case "name":
-// 		// 		if len(values) > 0 {
-// 		// 			filters = append(filters, filter{
-// 		// 				field: fmt.Sprintf("%s LIKE ?", fieldName),
-// 		// 				value: fmt.Sprintf("%%%s%%", values[0]),
-// 		// 			})
-// 		// 			// cond := fmt.Sprintf("%s LIKE ?", fieldName)
-// 		// 			// db.Where(cond, fmt.Sprintf("%%%s%%", values[0]))
-// 		// 		}
-// 		// 	}
-// 		// }
-// 		// if len(filters) > 0 {
-// 		// 	for i, f := range filters {
-// 		//
-// 		// 	}
-// 		// }
-// 		// return db
-// 	}
-// }
-
-// func temp(db *gorm.DB, filters []tempFilter) *gorm.DB {
-// 	for _, filter := range filters {
-// 		db.Or(filter.condition, filter.value)
-// 	}
-// }
-
-// func (r usersRepo) SearchFilter(fields url.Values) repo.QueryFilter {
-// 	return func(db *gorm.DB) *gorm.DB {
-// 		if len(fields) == 0 {
-// 			return db
-// 		}
-// 		// conditions := r.filterStrings(db, fields)
-// 		// d := db.Unscoped()
-// 		return db.Where(r.filterStrings(db, fields))
-// 	}
-// }
