+++ src/db.go
@@ -3,7 +3,6 @@ package goat
 import (
 	"gorm.io/gorm"
 
-	"github.com/68696c6c/goat/query"
 	"github.com/68696c6c/goat/sys/database"
 )
 
@@ -23,8 +22,4 @@ type DatabaseConnection database.ConnectionConfig
 
 func GetCustomDB(c DatabaseConnection) (*gorm.DB, error) {
 	return g.DB.GetConnection(database.ConnectionConfig(c))
-}
-
-func ApplyPaginationToQuery(q query.Builder, baseGormQuery *gorm.DB) error {
-	return g.DB.ApplyPaginationToQuery(q, baseGormQuery)
 }
