--- src/query/filter/logic.go
@@ -1,8 +1,0 @@
-package filter
-
-type logic string
-
-const (
-	logicAnd logic = "AND"
-	logicOr  logic = "OR"
-)
