--- src/request_validator.go
@@ -1,11 +1,0 @@
-package goat
-
-// import "gopkg.in/go-playground/validator.v8"
-//
-// type RequestValidator struct {
-// 	*validator.Validate
-// }
-//
-// func (v *RequestValidator) ValidateStruct(i any) error {
-// 	return v.Struct(i)
-// }
