+++ src/resource/message.go
@@ -1,0 +1,6 @@
+package resource
+
+type MessageResponse struct {
+	Message string `json:"message"`
+	Resource
+}
