+++ src/sys/app.go
@@ -2,17 +2,16 @@ package sys
 
 import (
 	db "github.com/68696c6c/goat/sys/database"
-	"github.com/68696c6c/goat/sys/http/router"
 	log "github.com/68696c6c/goat/sys/logging"
+	"github.com/68696c6c/goat/sys/router"
 )
 
 type Goat struct {
 	config    Config
 	HttpDebug bool
 	DB        db.Service
-	// HTTP      http.Service
-	Log    log.Service
-	Router router.Service
+	Log       log.Service
+	Router    router.Service
 }
 
 func Init() Goat {
@@ -21,8 +20,7 @@ func Init() Goat {
 		config:    config,
 		HttpDebug: config.HttpDebug,
 		DB:        db.NewService(config.DB),
-		// HTTP:      http.NewServiceGin(config.HTTP),
-		Log:    log.NewServiceLogrus(config.Log),
-		Router: router.NewService(config.Router),
+		Log:       log.NewService(config.Log),
+		Router:    router.NewService(config.Router),
 	}
 }
