+++ src/sys/logging/service.go
@@ -14,17 +14,20 @@ import (
 const (
 	loggerPath  = "logs"
 	loggerExt   = "log"
-	loggerFile  = "custom"
+	loggerFile  = "custom.log"
 	loggerLevel = logrus.InfoLevel
 )
+
+// TODO: replace with zap???
 
 // Default logging is done to standard out.
 // Custom loggers can be created using log files if needed, but Goat is
 // optimized for using a single logger writing to standard out for ease of use
 // with Docker and Amazon ECS.
+
 type Service interface {
 	NewLogger() *logrus.Logger
-	NewFileLogger(path string) (*logrus.Logger, error)
+	NewFileLogger(fileName ...string) (*logrus.Logger, error)
 	GetLogPath() string
 	GetLogExt() string
 	GetLogLevel() string
@@ -42,7 +45,7 @@ type service struct {
 	level      logrus.Level
 }
 
-func NewServiceLogrus(c Config) Service {
+func NewService(c Config) Service {
 	// If an invalid log level is provided, fallback to the default.
 	level, err := logrus.ParseLevel(c.Level)
 	if err != nil {
@@ -55,46 +58,46 @@ func NewServiceLogrus(c Config) Service {
 	}
 }
 
-// Returns the configured log file path.
+// GetLogPath returns the configured log file path.
 func (s service) GetLogPath() string {
 	return s.logPath
 }
 
-// Returns the configured log file extension.
+// GetLogExt returns the configured log file extension.
 func (s service) GetLogExt() string {
 	return s.logFileExt
 }
 
-// Returns the configured default log level.
+// GetLogLevel returns the configured default log level.
 func (s service) GetLogLevel() string {
 	return s.level.String()
 }
 
-// Returns a new standard out logger.
+// NewLogger returns a new logger instance that writes to standard out.
 func (s service) NewLogger() *logrus.Logger {
 	l := logrus.New()
 	l.SetLevel(s.level)
 	return l
 }
 
-// Returns a new logger that writes to the specified file, relative to the
-// configured log path using the configured log file extension.
-// If a file name is not specified a default name is used.
-func (s service) NewFileLogger(name string) (*logrus.Logger, error) {
+// NewFileLogger returns a new logger that writes to the specified file, relative to
+// the configured log path.If a file name is not specified a default name is used.
+func (s service) NewFileLogger(fileName ...string) (*logrus.Logger, error) {
 	logger := logrus.New()
 
-	if name == "" {
-		name = loggerFile
+	name := loggerFile
+	if len(fileName) > 0 && fileName[0] != "" {
+		name = fileName[0]
 	}
 
-	path := fmt.Sprintf("%s/%s.%s", s.logPath, name, s.logFileExt)
+	path := fmt.Sprintf("%s/%s", s.logPath, name)
 
-	fname, err := filepath.Abs(path)
+	logPath, err := filepath.Abs(path)
 	if err != nil {
 		return nil, errors.Wrap(err, "error resolving log file path")
 	}
 
-	file, err := os.OpenFile(fname, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0664)
+	file, err := os.OpenFile(logPath, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0664)
 	if err != nil {
 		return nil, errors.Wrap(err, "error opening log file")
 	}
