--- src/sys/utils/args.go
@@ -1,9 +1,0 @@
-package utils
-
-func ArgStringD(arg, def string) string {
-	v := arg
-	if arg == "" {
-		v = def
-	}
-	return v
-}
