+++ src/sys/utils/utils_test.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"testing"
 
+	"github.com/icrowley/fake"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -15,4 +16,17 @@ func Test_ValueOrDefault_Input(t *testing.T) {
 func Test_ValueOrDefault_Default(t *testing.T) {
 	result := ValueOrDefault[string](nil, "bar")
 	assert.Equal(t, "bar", result)
+}
+
+func TestArgStringD_Provided(t *testing.T) {
+	a := fake.Word()
+	d := fake.Word()
+	v := ArgStringD(a, d)
+	assert.Equal(t, a, v, "failed to return provided arg")
+}
+
+func TestArgStringD_Default(t *testing.T) {
+	d := fake.Word()
+	v := ArgStringD("", d)
+	assert.Equal(t, d, v, "failed to return default arg")
 }
